smartdoc_services/functions/requirements.txt
firebase-functions
firebase-admin
google-generativeai
requests
beautifulsoup4
